src/main/java/uk/gov/companieshouse/disqualifiedofficersdataapi/converter/DisqualifiedOfficerWriteConverter.java
package uk.gov.companieshouse.disqualifiedofficersdataapi.converter;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.mongodb.BasicDBObject;
import org.springframework.core.convert.converter.Converter;
import org.springframework.data.convert.WritingConverter;
import uk.gov.companieshouse.disqualifiedofficersdataapi.model.DisqualificationDocument;

@WritingConverter
public class DisqualifiedOfficerWriteConverter implements Converter<DisqualificationDocument, BasicDBObject> {

    private final ObjectMapper objectMapper;

    public DisqualifiedOfficerWriteConverter(ObjectMapper objectMapper) {
        this.objectMapper = objectMapper;
    }

    /**
     * Write convertor.
     * @param source source Document.
     * @return charge BSON object.
     */
    @Override
    public BasicDBObject convert(DisqualificationDocument source) {
        try {
            return BasicDBObject.parse(objectMapper.writeValueAsString(source));
        } catch (Exception ex) {
            throw new RuntimeException(ex);
        }
    }
}
